Convert all math to 32bit types. This improves overall accuracy and performance, it also simplifies the code considerably.

+++ includes/acl/io/clip_reader.h
@@ -255,7 +255,7 @@ namespace acl
 
 		bool create_clip(std::unique_ptr<AnimationClip, Deleter<AnimationClip>>& clip, const RigidSkeleton& skeleton)
 		{
-			clip = make_unique<AnimationClip>(m_allocator, m_allocator, skeleton, m_num_samples, m_sample_rate, String(m_allocator, m_clip_name), m_error_threshold);
+			clip = make_unique<AnimationClip>(m_allocator, m_allocator, skeleton, m_num_samples, m_sample_rate, String(m_allocator, m_clip_name), (float)m_error_threshold);
 			return true;
 		}
